Theingi emp save impl (#9)
* test

* Update Employee

* update some part

* Commit

+++ LapsApplication/src/main/java/sg/edu/iss/sa50/t8/model/Manager.java
@@ -14,7 +14,7 @@ public class Manager extends Staff{
 	//Bianca New constructors Jue 28th
 	public Manager(String name, String email,
 			Manager manager, 
-			int annualLeaveDays, int medicalLeaveDays) {
+			long annualLeaveDays, long medicalLeaveDays) {
 		super(name,email,manager,annualLeaveDays,medicalLeaveDays);
 	}
